restructured to have a reusable test for user defined archive modules
