+++ opt4j-core/src/integration-test/java/org/opt4j/core/common/archive/AbstractArchiveOptimalityTester.java
@@ -1,19 +1,28 @@
+package org.opt4j.core.common.archive;
+
 import java.util.Random;
 
 import org.junit.Assert;
-import org.junit.Test;
 import org.opt4j.core.Individual;
 import org.opt4j.core.Objective;
 import org.opt4j.core.Objectives;
-import org.opt4j.core.common.archive.ArchiveModule;
-import org.opt4j.core.common.archive.ArchiveModule.Type;
 import org.opt4j.core.optimizer.Archive;
+import org.opt4j.core.start.Opt4JModule;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Provider;
 
-public class ArchiveOptimalityTest {
+/**
+ * The {@link AbstractArchiveOptimalityTester} tests an Archive configured with the given {@link Opt4JModule} against
+ * different optimality criterions.
+ * <p>
+ * This class can be reused for additional {@link Archive}s to test.
+ * 
+ * @author Felix Reimann
+ *
+ */
+public class AbstractArchiveOptimalityTester {
 	private static final int NUMBER_OF_INDIVUDUALS = 20000;
 
 	protected final Objective o1 = new Objective("counter");
@@ -24,33 +33,13 @@ public class ArchiveOptimalityTest {
 	private Individual last;
 	private Individual randMin;
 
-	@Test
-	public void defaultArchive() {
-		archiveModuleTester(new ArchiveModule());
-	}
-
-	@Test
-	public void adaptiveGridArchive() {
-		ArchiveModule adaptiveGrid = new ArchiveModule();
-		adaptiveGrid.setType(Type.ADAPTIVE_GRID);
-		archiveModuleTester(adaptiveGrid);
-	}
-
-	@Test
-	public void crowdingArchive() {
-		ArchiveModule crowding = new ArchiveModule();
-		crowding.setType(Type.CROWDING);
-		archiveModuleTester(crowding);
-	}
-
-	@Test
-	public void unboundedArchive() {
-		ArchiveModule unbounded = new ArchiveModule();
-		unbounded.setType(Type.UNBOUNDED);
-		archiveModuleTester(unbounded);
-	}
-
-	protected void archiveModuleTester(ArchiveModule archiveModule) {
+	/**
+	 * Test a given {@link Opt4JModule}.
+	 * 
+	 * @param archiveModule
+	 *            the archive module under test
+	 */
+	protected void archiveOptimalityTest(Opt4JModule archiveModule) {
 		Injector injector = Guice.createInjector(archiveModule);
 		Archive archive = injector.getInstance(Archive.class);
 		System.out.println("Testing " + archive.getClass());
@@ -63,10 +52,17 @@ public class ArchiveOptimalityTest {
 		Assert.assertTrue(archive.contains(last));
 		Assert.assertTrue(archive.contains(randMin));
 
-		testOptimality(archive);
+		testOptimalityOfAllIndividuals(archive);
 	}
 
-	protected void testOptimality(Archive archive) {
+	/**
+	 * This tests fail if an individual is found in the {@link Archive} which is dominated by other {@link Individual}s
+	 * in the {@link Archive}.
+	 * 
+	 * @param archive
+	 *            the archive under test
+	 */
+	private void testOptimalityOfAllIndividuals(Archive archive) {
 		while (!archive.isEmpty()) {
 			Individual i = archive.iterator().next();
 			archive.remove(i);
@@ -85,7 +81,22 @@ public class ArchiveOptimalityTest {
 		}
 	}
 
-	protected void fillArchive(Injector injector, Archive archive) {
+	/**
+	 * Initializes the {@link Archive} with NUMBER_OF_INDIVIDUALS different {@link Individual}s.
+	 * <p>
+	 * Therefore. three {@link Objective}s are added:
+	 * <ul>
+	 * <li>the first dimension is incremented</li>
+	 * <li>the second dimension is decremented</li>
+	 * <li>the third dimension is set randomly with a static seed.</li>
+	 * </ul>
+	 * 
+	 * @param injector
+	 *            the injector
+	 * @param archive
+	 *            the archive under test
+	 */
+	private void fillArchive(Injector injector, Archive archive) {
 		Random r = new Random(100);
 		Provider<Individual> individuals = injector.getProvider(Individual.class);
